Move Error and Statistics callbacks under ConsumerGroups namespace
Move `Karafka::Instrumentation::Callbacks::Error` and
`Karafka::Instrumentation::Callbacks::Statistics` to
`Karafka::Instrumentation::Callbacks::ConsumerGroups::Error` and
`Karafka::Instrumentation::Callbacks::ConsumerGroups::Statistics`.

Both callbacks carry consumer_group_id in their payload and are
registered per consumer-group connection client. This completes the
callbacks layer namespacing — all CG-specific librdkafka callbacks
now live under Callbacks::ConsumerGroups alongside Rebalance.

Published event names and payload shape are unchanged.
Internal-only move, no alias needed.
