--- lib/karafka/instrumentation/callbacks/error.rb
@@ -1,54 +1,0 @@
-# frozen_string_literal: true
-
-module Karafka
-  module Instrumentation
-    # Callbacks used to transport things from rdkafka
-    module Callbacks
-      # Callback that kicks in when consumer error occurs and is published in a background thread
-      class Error
-        include Helpers::ConfigImporter.new(
-          monitor: %i[monitor]
-        )
-
-        # @param subscription_group_id [String]
-        # @param group_id [String] id of the owning group (consumer group today)
-        # @param client_name [String] rdkafka client name
-        def initialize(subscription_group_id, group_id, client_name)
-          @subscription_group_id = subscription_group_id
-          @group_id = group_id
-          @client_name = client_name
-        end
-
-        # Runs the instrumentation monitor with error
-        # @param client_name [String] rdkafka client name
-        # @param error [Rdkafka::Error] error that occurred
-        # @note It will only instrument on errors of the client of our consumer
-        def call(client_name, error)
-          # Emit only errors related to our client
-          # Same as with statistics (mor explanation there)
-          return unless @client_name == client_name
-
-          monitor.instrument(
-            "error.occurred",
-            caller: self,
-            subscription_group_id: @subscription_group_id,
-            consumer_group_id: @group_id,
-            group_id: @group_id,
-            type: "librdkafka.error",
-            error: error
-          )
-        rescue => e
-          monitor.instrument(
-            "error.occurred",
-            caller: self,
-            subscription_group_id: @subscription_group_id,
-            consumer_group_id: @group_id,
-            group_id: @group_id,
-            type: "callbacks.error.error",
-            error: e
-          )
-        end
-      end
-    end
-  end
-end
