+++ spec/lib/karafka/instrumentation/callbacks/consumer_groups/error_spec.rb
