+++ spec/lib/karafka/instrumentation/callbacks/consumer_groups/statistics_spec.rb
